feat: modal opens onclick of mentor, to show profile info
